add Settings page to desktop navigation
